Refactor: ToFloat() float 반환 + ToDouble() double 반환 분리
- ToFloat(): double → float (32-bit) 반환 타입 변경
- ToDouble(): double (64-bit) 유지
- MathModule/StatisticsModule: ToFloat() → ToDouble() (정밀도 보존)
- PyInt 중복 ToDouble() 제거

+++ core/PyObject.cs
@@ -771,9 +771,9 @@ namespace SharpPy
         }
 
         /// <summary>
-        /// CPython PyFloat_AsDouble 호환: PyObject에서 C# double 값 추출
+        /// CPython PyFloat_AsDouble 호환: PyObject에서 C# float (32-bit) 값 추출
         /// </summary>
-        public virtual double ToFloat()
+        public virtual float ToFloat()
         {
             throw PyTypeError.Create($"float() argument must be a string or a number, not '{GetTypeName()}'");
         }
