+++ type/PyIntSubclass.cs
@@ -335,7 +335,7 @@ namespace SharpPy
         #region Type Conversion - Delegate to PyInt
 
         public override int ToInt() => _intValue.ToInt();
-        public override double ToFloat() => _intValue.ToFloat();
+        public override float ToFloat() => _intValue.ToFloat();
         public override double ToDouble() => _intValue.ToDouble();
         public override bool PyBoolValue() => _intValue.PyBoolValue();
         public override PyString ToStr() => _intValue.ToStr();
